update validator provider
