+++ packages/ui/src/validators/providers/context.tsx
@@ -2,4 +2,7 @@ import { createContext } from 'react'
 
 import { UseValidators } from './provider'
 
-export const ValidatorsContext = createContext<UseValidators>({ setShouldFetchValidators: () => {} })
+export const ValidatorsContext = createContext<UseValidators>({
+  setShouldFetchValidators: () => {},
+  setShouldFetchExtraDetails: () => {},
+})
